task 3(d) and 4 (25/04/2020)
task 3(d) and 4 (25/04/2020). Declare the essence and describe its properties. 
Declare entity arrays. Fill attributes with numerical and text values (define the range of values yourself).
Analyze the dependence of the number of permutations on the number of elements in an array.
For each of the arrays, implement a key (structure attribute) search.
For each of the arrays to implement sorting one key.